refactor and better signature
