+++ mobile/lib/utils/version_compatibility.dart
@@ -1,17 +1,17 @@
-(bool, String, int) checkVersionCompatibility(
+String? getVersionCompatibilityMessage(
   int appMajor,
   int appMinor,
   int serverMajor,
   int serverMinor,
 ) {
   if (serverMajor != appMajor) {
-    return (false, 'major', appMajor);
+    return 'Your app major version is not compatible with the server!';
   }
 
   // Add latest compat info up top
   if (serverMinor < 106 && appMinor >= 106) {
-    return (false, 'minor', 106);
+    return 'Your app minor version is not compatible with the server! Please update your server to version v1.106.0 or newer to login';
   }
 
-  return (true, '', 0);
+  return null;
 }
